extract types to own file
