+++ src/cards/a-card.ts
@@ -1,4 +1,5 @@
-import { Card, CardInstance, addCard } from "../cardContainer"
+import { Card, CardInstance } from "../types"
+import { addCard } from "../cardContainer"
 
 const acrobatic: Card = {
   name: 'ACROBATIC',
